Add files via upload
This key is what your application use for encryption and decryption which you should keep secret else anyone can decrypt your data with that key and it must be either 128 bit or 192 bit (not even in between). If you don’t know that your supplied key is either 128/192 bit or not you can bit calculation.